capstone-projects/security-analyst-copilot/requirements.txt
# Core
langchain>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.20

# UI
streamlit>=1.30.0
rich>=13.0.0

# Data
chromadb>=0.4.0
pandas>=2.0.0

# Integrations
requests>=2.31.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
